requirements.txt
griffe
pydantic
pydantic-ai
fastapi
uvicorn[standard]
python-multipart
python-dotenv
google-generativeai
llama-index
llama-index-embeddings-huggingface
sentence-transformers
torch
transformers
huggingface_hub